TDS_Projet_python/requirements.txt
appdirs==1.4.4
asn1crypto==0.24.0
astroid==1.6.6
autopep8==1.5
backports.functools-lru-cache==1.6.1
Click==7.0
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.1.4
distlib==0.3.1
Django==1.11.29
django-rest-framework==0.1.0
djangorestframework==3.9.4
enum34==1.1.10
filelock==3.0.12
Flask==1.1.1
futures==3.3.0
idna==2.6
importlib-metadata==2.1.0
importlib-resources==3.3.0
ipaddress==1.0.17
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.6
opencv-python==4.2.0.32
pathlib2==2.3.5
pep8==1.7.1
Pillow==6.2.2
pycodestyle==2.5.0
pycrypto==2.6.1
pygobject==3.26.1
pylint==1.9.5
pytesseract==0.3.6
pytz==2019.3
pyxdg==0.25
redis==3.4.1
scandir==1.10.0
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.15.0
tesseract==0.1.3
typing==3.7.4.3
virtualenv==20.2.1
Werkzeug==1.0.0
wrapt==1.12.1
zipp==1.2.0
